+++ NEWS.md
@@ -1,11 +1,13 @@
-Version 0.2.11.999:  May 3, 2019
+Version 0.2.11.999:  May 8, 2019
 -------------------------------------------------------------------------------
 
 General:
 
++ Added `ape` and `tibble` dependencies.
+
 Lineage:
 
-+ Added `readIgphyml` to read in IgPhyML output, and `combineIgphyml` to 
++ Added `readIgphyml` to read in IgPhyML output and `combineIgphyml` to 
   combine parameter estimates across samples.
 
 Diversity: